Refactored package structure for better organization
