+++ automated-judge-system/src/main/java/com/voidlings/AssignmentHandling/AssignmentClass.java
@@ -1,4 +1,4 @@
-package com.voidlings;
+package com.voidlings.AssignmentHandling;
 
 public class AssignmentClass{
