+++ automated-judge-system/src/main/java/com/voidlings/AssignmentHandling/AssignmentMethod.java
@@ -1,4 +1,5 @@
-package com.voidlings;
+package com.voidlings.AssignmentHandling;
+
 public class AssignmentMethod{
 
 }
