+++ automated-judge-system/src/main/java/com/voidlings/EvaluationHandling/ClassEvaluation.java
@@ -1,4 +1,4 @@
-package com.voidlings;
+package com.voidlings.EvaluationHandling;
 
 public class ClassEvaluation implements EvaluationStrategy{
