+++ automated-judge-system/src/main/java/com/voidlings/EvaluationHandling/EvaluationStrategy.java
@@ -1,4 +1,4 @@
-package com.voidlings;
+package com.voidlings.EvaluationHandling;
 
 public interface EvaluationStrategy {
