+++ automated-judge-system/src/main/java/com/voidlings/FileHandling/Folder.java
@@ -1,4 +1,4 @@
-package com.voidlings;
+package com.voidlings.FileHandling;
 
 import java.io.File;
 import java.util.ArrayList;
