+++ automated-judge-system/src/main/java/com/voidlings/FileHandling/JavaFile.java
@@ -1,4 +1,4 @@
-package com.voidlings;
+package com.voidlings.FileHandling;
 import java.util.ArrayList;
 import java.util.List;
 import java.io.File;
