+++ automated-judge-system/src/main/java/com/voidlings/FileHandling/ZIPFileHandler.java
@@ -1,4 +1,4 @@
-package com.voidlings;
+package com.voidlings.FileHandling;
 import java.io.*;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
