+++ automated-judge-system/src/main/java/com/voidlings/ReportHandling/EvaluationReport.java
@@ -1,4 +1,4 @@
-package com.voidlings;
+package com.voidlings.ReportHandling;
 
 public class EvaluationReport {
