+++ automated-judge-system/src/main/java/com/voidlings/SpecificationHandling/AssignmentAttributeSpec.java
@@ -1,4 +1,7 @@
-package com.voidlings;
+package com.voidlings.SpecificationHandling;
+
+import com.voidlings.AssignmentHandling.AssignmentAttribute;
+
 
 public class AssignmentAttributeSpec extends AssignmentAttribute{
