+++ automated-judge-system/src/main/java/com/voidlings/SpecificationHandling/AssignmentClassSpec.java
@@ -1,4 +1,6 @@
-package com.voidlings;
+package com.voidlings.SpecificationHandling;
+
+import com.voidlings.AssignmentHandling.AssignmentClass;
 
 public class AssignmentClassSpec extends AssignmentClass{
