+++ automated-judge-system/src/main/java/com/voidlings/SpecificationHandling/AssignmentClassSpecBuilder.java
@@ -1,4 +1,4 @@
-package com.voidlings;
+package com.voidlings.SpecificationHandling;
 
 public class AssignmentClassSpecBuilder implements AssignmentComponentBuilder{
