+++ automated-judge-system/src/main/java/com/voidlings/SpecificationHandling/AssignmentMethodSpec.java
@@ -1,4 +1,6 @@
-package com.voidlings;
+package com.voidlings.SpecificationHandling;
+
+import com.voidlings.AssignmentHandling.AssignmentMethod;
 
 public class AssignmentMethodSpec extends AssignmentMethod{
