+++ automated-judge-system/src/main/java/com/voidlings/SpecificationHandling/SpecDirector.java
@@ -1,4 +1,4 @@
-package com.voidlings;
+package com.voidlings.SpecificationHandling;
 
 public class SpecDirector{
