+++ automated-judge-system/src/main/java/com/voidlings/TestCases/AttributeTestCase.java
@@ -1,4 +1,4 @@
-package com.voidlings;
+package com.voidlings.TestCases;
 
 public class AttributeTestCase implements TestCase{
