+++ automated-judge-system/src/main/java/com/voidlings/TestCases/AttributeTestCaseFactory.java
@@ -1,4 +1,4 @@
-package com.voidlings;
+package com.voidlings.TestCases;
 
 public interface AttributeTestCaseFactory extends TestCaseFactory {
