+++ automated-judge-system/src/main/java/com/voidlings/TestCases/MethodTestCase.java
@@ -1,4 +1,4 @@
-package com.voidlings;
+package com.voidlings.TestCases;
 
 public class MethodTestCase implements TestCase{
