+++ automated-judge-system/src/main/java/com/voidlings/TestCases/MethodTestCaseFactory.java
@@ -1,4 +1,4 @@
-package com.voidlings;
+package com.voidlings.TestCases;
 
 public interface MethodTestCaseFactory extends TestCaseFactory{
